Fix for package import
Previous attempt to implement import of Murano packages was buggy.
It didn't assume that names in packages/application_package and in
db/models are different. We need to map names from one format to
another.

A big piece of functionality introduced in this patch is related
to archiving of package definition in a zip file. That is required
by our package format stored in the database.

Change-Id: I9660ba6fde00d777ec332c4912c972d1e2744718