Configure rabbit vhost for Murano
MURANO_RABBIT_VHOST allows to specify a separate virtual host for Murano services.

This is not required if all OpenStack services are deployed by devstack scripts
  on a single node. In this case '/' virtual host (which is the default) is enough.

The problem arise when Murano installed in 'devbox' mode, allowing two or more
  devboxes to use one common OpenStack host. In this case it's better devboxes
  use separated virtual hosts, to avoid conflicts between Murano services.

This couldn't be done using exitsting variables, so that's why this variable was added.

Change-Id: Ib9cfc6b23afc0362226b3e2a891befba7595f776