Use HOT instead of CFN heat template format
New Heaet features will get added to HOT but possibly not the
older CFN format, and it's easier to make the change now than
later.

The only code change is in heat_stack to look for 'resources'
instead of 'Resources'; all other changes are to the Heat template
sections within PL class definitions. Keys are lower-cased and
in some cases changed to match HOT (e.g. key_name instead of KeyName),
and sections specifying Heat resources add heat_template_version to
the template.

Implements blueprint replace-cfn-templates-with-hot

Change-Id: If4d34cd25a4e964d7555fedf218512d5c0f5615c