Make environment name unique within a tenant
Duplicate environment names cause all kinds of problems because of
naming conventions used. There's no real reason to allow them (heat
allows duplicate stack names but there was a lot of discussion about
it, and the reason for having them was nobody wanted to remove them).
In our case, duplicate environment names cause trouble and there is
no advantage to them.

Server will return a 409 Conflict in this instance.

Change-Id: Ibe694c3dab19d77f6de3aa31844ff3e3cc0d9eb4
Closes-Bug: 1314106