Move Neutron networking implementation to Core Library
To support future advanced scenarios (such as SharedIP/Cluster groups) it is required to give the Instance
ownership of its port. This requires to move appropriate generation of Heat snippet from Network class
to Instance class.
This makes murano.lib.networks.Neutron package useless, as Neutron-specific leaks to CoreLibrary anyway.
So, it's better to merge these two libs

Also, introduces a concept of "primary network" for instance: this is the net, in which the
floating ip (and other advanced stuff) is allocated.

Change-Id: Ie7eca72901e679707fd63176ced99167eb428646
Targets-blueprint: neutron-net-to-murano-core