Introduce a SharedIp object for Clustering
An io.murano.SharedIp class is added to the Core library.
This class may be used to allocate an IP address which will be later 
assigned to multiple VMs (with usage of "allowed_address_pairs") feature.
A FloatingIP address may be attached to this shared virtual address for further usage from outer world.

Also fixes an issue in NeutronNetwork which caused the net to be redeployed multiple times.

Change-Id: If821348d27c7cd8d90fe11a7f9c9fe9e0e558c03