Handle null JSON blobs
The JSON deserializer doesn't handle null values properly; it raises
a TypeError because it expects a string or buffer. This became apparent
after the supplier info update which added columns that would be null
for existing packages. Patch returns None from the JSON deserializer
in that case, and also refactors unittest base classes.

Change-Id: I6f31500d2dd1aaaf6a1fd26c1901666072d79776
Closes-Bug: #1342306