Make apache restarts a little more forgiving
Apache sometimes is slow to release the port which can cause restarts of apache
to fail due to the requested port already being bound. Make use of the function
'restart_apache_server' added to fix #1342256, which introduces a small sleep
between the stop and start to help make apache deployments a bit more resilient.

Closes-Bug: #1358281

Change-Id: Ic4206f24d7bb04b014b3cb179d2303bb301c59d2