Don't rely on OS::Nova::Server.addresses
Server.addresses's implementation in Heat is dependent on a nova extension
which is not guaranteed to be present. I'm not sure if this is something
that should be fixed in Heat or not, but this is a workaround that
results in the same information (instance.ipAddresses will be set to a
list of all IP addresses across all networks the instance is assigned
to).

Change-Id: I5154f7e2a69ab8e64f9fbd1d8a26daf07f2482cd
Closes-Bug: 1364446