Extract version definition to a separate file
Before, version was stores in the init file.
It's better not to store such kind of public data
as version to init file.
Moreover, whole murano package needed to be imported to identify project version
Now, separate file is used for that

Also, doc builder config was updated to support this change

Change-Id: Iede1dfbe9ae4d8c6fb5ecdeb8f8f25124e4bc34f