Remove hardcoded dns from config file
If no nameserver list is provided to a neutron subnet,
it uses the DHCP (gateway) address which in turn forwards to dnsmasq.

So dns_nameservers will be set to emptylist in heat templete.

In case user wants to set different dns from default,
he may edit the config parameter, witch is not removed.
The name of parameters is changes to dns_nameservers (with s at the end)
to correspond the same name in heat template.

Change-Id: Iba8b12b8bc78f821e11debb58f20cb713a9dc9c7
Closes-Bug: #1367877