Fix for HOT parameter types conversion
The HOT-to-MuranoPL converter did not support some types for Heat
Template's input parameters.
Type 'comma_delimited_list' was not supported at all which was leading
to unhandled exceptions during package import operation.
Type 'json' was suported, but the generated UI definition did not
contain 'type' attribute for appropriate field, which was leading to
unhadled exceptions when the UI form was about to be displayed.

This patch fixes both these issues.

Closes-Bug: #1450536

Change-Id: I539514cc7cd884f98278cf5c990e8833e1b09259