Update from global-requirements.txt and fix cli tests
Commit ID from openstack/requirements:
40b86ac9cb707ff97f5b140d4f5fe5f1dcbbb2ed

There are a few changes in how CLI client is setup in functional
tests after tempest-lib updated to 0.5. A temporary workaround
is introduced to overcome lack of current configuration of CLI
client tests. Next step - move CLI tests to python-muranoclient
since this is a common direction in community.

Change-Id: I059be13fc0af4a0faec917322b4f82d0f4a81117