Make use of devstack external plugin
Instead of copying files manually to devstack it much
easier to use the devstack external plugin interface [1].
This makes the usage of Murano much easier for developers.

The plugin.sh script is a copy of contrib/devstack/lib/*
and contrib/devstack/extras.d/70-murano.sh. These scripts
can be removed after the CI system is switch to the new
model.

[1]: http://docs.openstack.org/developer/devstack/plugins.html#externally-hosted-plugins

Change-Id: I9fc24d17301947f5a95c139576037efc679d4bc4
Implements: bp use-devstack-plugin