Added script for unit tests coverage job
This script will allow to run non voiting coverage job which will detect
any changes in the code coverage for each commit on review.
It will help to automatically control the unit tests coverage for each commit
and core team can ask to add unit tests to some commit which will require it.

Change-Id: I318a6b0133333ef16494ed2def512720b15ce956
Closes-Bug: #1473364