Use Custom networks as primary if no others exist
Modify the logic of primary network detection, so if the instance does
not join environment's default of flat networks, but some custom
networks are passed, the first of them is used as a primary one.

This fixes the issue when the floating ip address could not be assigned
in case of custom networking configuration.

Change-Id: Ief8819fd24695aceb118862839301b8045c41ea4
Closes-bug: #1474092