Prevent 500 when requesting deployments of a deleting env
Currently a 500 error would be raised if we request list of deployments
of an env, that is being deleted. This due to the helper function, that
expects description of an environment to be a dict.
This patch makes the helper function aware, that description can be
None, therefore preventing the 500 error.

Change-Id: I9ee05dcb583fdca59813b23d948ec4baa20b713b
Closes-Bug: #1481840