Glance MuranoPackage Artifact
Adds an artifact type for Glance V3 to define murano package, its
specific properties and binary objects (full package, logo and a ui
definition).

As a temporary workaround to simplify search a 'keywords' field is added
to contain display name and tags word-by-word, so any of them may be
used to locate a package. This will be inefficient in terms of DB
performance, but at least not worse then the legacy solution.

Implements-blueprint: artifact-repository-support
Change-Id: I11efa3758cea5eb459a101d62c93355f34bb42f9