Enforced AUTO replacement_policy for sharedIp port
In Openstack releases prior to Kilo Heat recreates all the Neutron Ports
during the stack update regardless of the port resource being actually
updated or not. This is done intentionally to workaround the bug in nova
(https://bugs.launchpad.net/nova/+bug/1158684)
However, this approach may cause issues when the stack is updated after
the instance and the port are created (which happens during many Murano
env deployments). To control port re-creation logic Heat allows to
explicitly specify the replacement_policy property for OS::Neutron::Port
resource: if set to AUTO it forbids Heat to recreate the port if it was
not modified.

For regular instances and their ports we are already using this setting.
However, for the ports created for SharedIps this was not true, thus
causing issues on pre-kilo deployments. This patch fixes it.

On deployments using Kilo+ Heat this patch has no effect, as the default
value of replacenent_policy is already AUTO in this releases.

Change-Id: Ib87961a4f2285bf622df50828ecd2ad98365a910
Closes-Bug: #1488105