Namespace resolution error was fixed
For method invocation context of a sender object was used.
Then the current type was taken from it and class names were
resolved using namespaces of that class. However in situations
when contracted method was located in parent class that had
different namespace declarations it lead to a wrong namespaces
being used.

This commit adds information on which class has the contract
so that the sender could be upcasted to it first.

Change-Id: Ieb4fca4ea7f2c64c7a731b81fee2ccaff1b1a531
Closes-Bug: #1489524