Fixed incorrect MuranoPL names for some of Python-based methods
When method that was defined in Python get registered in MuranoPL
it name changes according to yaql conventions (and thus report_error
becomes reportError). However this translation was buggy because
1. It didn't respect dsl.name('name') decorator that was supposed to
override default behavior
2. It didn't strip underscores from the right of the name as yaql usually
does (that prevent from name collisions in Python)

Because of those issues applications that relied on StatusReporter::report_error
method became broken.

This commit fixes both of the issues with name translation.

Change-Id: I84d4bd4df1af5bbe150f89c096afd8c52ad3da1a
Closes-Bug: #1493875