Function caller was incorrect when called from Parallel block
Current instruction is tracked as a value in yaql context.
But in Parallel block individual instructions executed in separate
threads. As a result when those threads get scheduled current
instruction in caller context already changes to the last one
(which is a synchronization point for the Parallel block).

To solve this all spawned threads need to work in their own
dedicated child contexts. Then each of those contexts will have
its own "current instruction". However this also brings a side effect:
all local variable modifications made by spawned threads will
be go to those new contexts and will be lost upon exit from Parallel.
To solve this context that local variable need to be in is marked with
special flag. When executing "$var: value" expressions DSL will scan
context chain up to the context with that flag. Currently the only
context with the flag is a method context that defines local variable scope.

Change-Id: I0647915c104dfe4e00dd86c17726245d1bbc3beb
Closes-Bug: #1497772