Specification of which property/argument violated contract was added
When some property or argument value violated its contract it was impossible
to tell which property/argument caused the exception.

This change adds prefix to ContractViolationException message that tells
not only which property or argument failed to validate but for composite
properties also the path within the data.
For example if we change ApacheHttpServer name property contract from
$.string() to $.int() we will get

[io.murano.Environment.applications[0]]
[io.murano.apps.apache.ApacheHttpServer.name]
Value 'ApacheHttpServer' violates int() contract

Also for check() contracts ability to provide custom
error message was added

Change-Id: I6953ec84140f4bed3d50aa181244f89682b2b2fb
Closes-Bug: #1496044