Do not try to delete server from deleted Heat stack
Because now Environment's .destroy method would
be called before that of an Instance the later will try
to delete itself from the stack that is physically not
existing and represented as an empty dictionary in
HeatStack.current().

With this change Instance will check stack existence
before deleting anything.

Also a method to check if Instance is present in Heat
stack template was added so that application could check
if it is safe to use the instance (for example to send Agent plans)

Change-Id: I298f3f079a153ef87ed7213b7f1432c9e5e4b4b6
Closes-Bug: #1503715