Increased the number of environments per tenant
The 'max_environments' setting of [networking] group in murano
configuration file defines the maximum number of networks which may be
created by murano for any given router, thus eventually limiting the
number of environments to simultaneously co-exists within a tenant.

The previous default (20) was very low, and it was causing CIDR
conflicts even when the actual number of envs was not reaching the
limit.

This change increases the number of CIDRs allowed for environment
networks, thus reducing the probability of CIDR conflict.

NOTE: This change just reduces the risk of conflicts but does not
eleminate it completely.

Change-Id: Id913d17b8f7207afc9b1983287349a6d70a09edf
Partial-bug: #1502437