Improve message for publish package conflict
Previously, end user will receive a conflict message "Class with the same full
name is already registered in the visibility scope" when publish package
conflict with one public package with same fqn. The "class" concept might be
not familiar to end user who doesn't know murano package details.

And in most cases, user will modify the class name in package. So it would be
better to check the duplicate fqn at first, which it give back a more straight
message "Package with the same Name is already made public" in this case.

Change-Id: I721ce60cb213fc446f62755b3b59523aca5d5fcb
Closes-Bug: #1506703