Add functional api tests for categories
- Added positive and negative tempest api tests for work with
categories
- Change inheritance: made separate class for authorisation, made
separate class that will contain methods for object creation (and
cleanup for them). Lately all methods for object creation will be
moved to this class and will use right cleanup mechanism (instead of
current implementation when objects save in array). This change was
made in this commit to avoid using incorrect solution in new tests.

Implements: blueprint tempest-tests-for-categories-management

Change-Id: I1fe71e385ab62c1d9fb9a9d90dd44f453aa953a9