HOT outputs were merged
This commit does dot HOT template outputs (for Heat.HOT/1.0 format)
the same that was previously done for the inputs. Instead of generating
MuranoPL property per each output now all outputs are put into a
single 'templateOutputs' property. This allows to eliminate the constraints
on output names, not to have collisions with input names and remain
consistent with previous changes.

Change-Id: Ibbd0a2f410c55385cd3cb9e85ea135a36460afe7
Closes-Bug: #1520954