Support for Cinder volumes was added
This change adds core library classes necessary to work with
cinder volumes:

Volume is the base class for cinder volume implementations.
It defines the interface for volumes. The most important methods of it
are "deploy" to deploy the volume and "attachTo" that returns a HOT
snippet to attach the volume to Instance

CinderVolume represents a new volume.
The class wraps OS::Cinder::Volume Heat resource.

ExistingCinderVolume is a Volume implementation for volumes
created outside of Murano.

CinderVolumeBackup and CinderVolumeSnapshot are
adapter classes to wrap around Cinder backups and snapshots
using their ID.

In addition new property "volumes" was added to the Instance class.
It is a mapping of a mounting path to Volume implementations
that the Instance must be attached to. Because the mapping
can be empty (or omitted) new property doesn't break backward
compatibility.

This commit doesn't has boot from volume functionality that need to be
designed separately.

Implements blueprint: cinder-volumes-support

Change-Id: I8ff7f8e4b48e7fb8112271a642cac703db0963aa